projects/claude_generated/python/p_33/p_33_requirements_claude.txt
# Spell checker libraries (install at least one)
pyspellchecker==0.8.1
textblob==0.18.0.post0

# TextBlob dependencies
nltk==3.9.1
